Moved Domino.cs to scripts
